+++ packages/next-sdk/page-tools/a11y/vnode.ts
@@ -1,0 +1,149 @@
+/**
+ * a11y/vnode.ts
+ *
+ * 处理无障碍树节点 (VNode) 的构建、剪枝与序列化逻辑。
+ */
+
+import { computeAccessibleName } from 'dom-accessibility-api'
+import { isFocusable } from 'tabbable'
+import type { VNode, RefMap, A11yTreeOptions } from './types'
+import { isHidden, inferRole, getStateTokens, collectDescendantText, getComposedChildren } from './utils'
+
+/**
+ * 递归将 DOM 元素转换为 VNode 中间表示
+ * @param el 当前 DOM 元素
+ * @param refCounter 引用计数器（使用对象引用避免全局可变状态）
+ * @param refMap ref 索引 → 元素映射
+ * @param blacklistSet 用户自定义黑名单
+ * @param whitelistSet 用户自定义白名单
+ * @param exposedAttributes 需要暴露的自定义属性
+ * @param errorSelectors 错误选择器
+ * @param warningSelectors 警告选择器
+ */
+export function buildVNode(
+  el: Element,
+  refCounter: { value: number },
+  refMap: RefMap,
+  blacklistSet: Set<Element>,
+  whitelistSet: Set<Element>,
+  exposedAttributes?: string[],
+  errorSelectors?: string,
+  warningSelectors?: string,
+): VNode | null {
+  if (isHidden(el) || blacklistSet.has(el)) return null
+
+  const role = inferRole(el)
+  const tokens = getStateTokens(el, exposedAttributes, errorSelectors, warningSelectors)
+  
+  let name = computeAccessibleName(el as HTMLElement)
+  const isTrulyInteractive = isFocusable(el as HTMLElement)
+  const isVisuallyClickable = tokens.includes('cursor=pointer')
+  // 包含白名单属性的节点也视为白名单节点（确保不被剪枝并分配 ref 操作索引）
+  const isWhitelisted = whitelistSet.has(el) || (exposedAttributes?.some(attr => el.hasAttribute(attr)) ?? false)
+  // <label for="..."> 原生可点击：浏览器将点击转发到关联的表单控件（checkbox/radio 等）
+  // Angular/React 自定义组件常隐藏原生 input，仅暴露 label 文本和自定义 skin
+  const isLabelFor = el.tagName.toLowerCase() === 'label' && el.hasAttribute('for')
+
+  // 兜底方案：如果 AccName 计算结果为空，但该节点具有明显的交互性或属于结构性列表项，
+  // 我们从其子树收集纯文本作为其 fallback name，以防 AI 丢失可读上下文。
+  // 注意：不要对 <select> / combobox / listbox 等下拉组件进行文本兜底，以防它们吸收所有子选项文本导致极其嘈杂。
+  if (!name.trim()) {
+    const tag = el.tagName.toLowerCase()
+    const isDropdown = tag === 'select' || role === 'combobox' || role === 'listbox'
+    if (!isDropdown) {
+      const isInteractiveTag = ['button', 'a', 'input', 'textarea', 'li', 'label'].includes(tag)
+      if (isTrulyInteractive || isWhitelisted || isVisuallyClickable || isInteractiveTag || role === 'listitem' || role === 'option') {
+        name = collectDescendantText(el)
+      }
+    }
+  }
+
+  // 最终兜底：非交互元素（如错误提示 ti-error-msg、状态信息）的纯文本捕获。
+  // 仅当元素没有有价值的子元素时触发，避免与子节点文本重复：
+  // - 叶子 span（只有文本节点）→ 捕获文本作为 name
+  // - 容器元素（有 element 子节点）→ 不触发，由子节点各自捕获
+  if (!name.trim()) {
+    const hasElementChildren = getComposedChildren(el).length > 0
+    if (!hasElementChildren) {
+      const text = collectDescendantText(el)
+      if (text) {
+        name = text
+      }
+    }
+  }
+
+  // generic 无 name 时，即使有 cursor=pointer 也不分配 ref：
+  // cursor 通常是 CSS 继承传播的，这类 div 本身无法被有意义地操作
+  const interactive = isTrulyInteractive || isWhitelisted || isLabelFor || (isVisuallyClickable && (role !== 'generic' || name !== ''))
+
+  let ref: number | undefined
+  if (interactive) {
+    ref = refCounter.value
+    refMap.set(ref, el as HTMLElement)
+    refCounter.value++
+  }
+
+  const children: VNode[] = []
+  for (const child of getComposedChildren(el)) {
+    const childVNode = buildVNode(child, refCounter, refMap, blacklistSet, whitelistSet, exposedAttributes, errorSelectors, warningSelectors)
+    if (childVNode) children.push(childVNode)
+  }
+
+  return { role, name, tokens, ref, el: el as HTMLElement, children }
+}
+
+/**
+ * 判断 VNode 子树是否包含任何有价值的节点（有 ref 或有 accessible name）
+ * 用于过滤空容器子树，避免输出无内容的嵌套层级
+ */
+export function hasValue(vnode: VNode): boolean {
+  if (vnode.ref !== undefined || vnode.name.trim() !== '') return true
+  return vnode.children.some(hasValue)
+}
+
+/**
+ * 判断节点是否需要透明穿透（跳过本节点但保留子节点）
+ *
+ * 统一规则：无 ref（非交互）且无 accessible name → 穿透
+ * 这样可以同时：
+ * - 去掉 generic/list/listitem 等纯布局噪音（无 name 时穿透）
+ * - 保留有 name 的 listitem（有 name 时保留，兼顾内容理解场景）
+ */
+export function shouldPassThrough(vnode: VNode, opts: Required<A11yTreeOptions>): boolean {
+  if (!opts.pruneUnnamed) return false
+  // preserveRoles 中的角色强制保留
+  if (opts.preserveRoles.includes(vnode.role)) return false
+  // 有 ref（交互节点）→ 永远保留
+  if (vnode.ref !== undefined) return false
+  // 有 accessible name → 保留（兼顾内容理解场景）
+  if (vnode.name.trim() !== '') return false
+  // 无 ref 且无 name → 透明穿透
+  return true
+}
+
+/**
+ * 将 VNode 序列化为 YAML 行数组
+ * 穿透节点时，子节点在当前 depth 平铺输出（不增加缩进层级）
+ */
+export function serializeVNode(
+  vnode: VNode,
+  depth: number,
+  opts: Required<A11yTreeOptions>,
+): string[] {
+  if (shouldPassThrough(vnode, opts)) {
+    // 透明穿透：跳过本节点，子节点保持当前 depth（层级不增加）
+    // 同时过滤掉整棵子树都无价值的空容器，避免输出无意义的嵌套
+    return vnode.children
+      .filter(c => hasValue(c))
+      .flatMap(c => serializeVNode(c, depth, opts))
+  }
+
+  const indent = '  '.repeat(depth)
+  const refStr = vnode.ref !== undefined ? ` #${vnode.ref}` : ''
+  const tokenStr = vnode.tokens.length > 0 ? ` [${vnode.tokens.join(' ')}]` : ''
+  const nameStr = vnode.name ? ` "${vnode.name}"` : ''
+  const line = `${indent}- ${vnode.role}${refStr}${tokenStr}${nameStr}`
+
+  const childLines = vnode.children.flatMap(c => serializeVNode(c, depth + 1, opts))
+  return [line, ...childLines]
+}
