+++ packages/next-sdk/page-tools/handlers/browserState.ts
@@ -1,0 +1,8 @@
+import type { ActionContext } from '../context'
+
+export async function handleBrowserState(args: any, ctx: ActionContext) {
+  if (window.__webmcpcli_beforeGetBrowserState) {
+    window.__webmcpcli_beforeGetBrowserState()
+  }
+  return ctx.buildBrowserStateResponse(args.responseMode ?? 'diff')
+}
